Fix integer overflow and type safety issues across giflib
Multiple integer overflow vulnerabilities exist in giflib's pixel count
and dimension arithmetic.  On 32-bit platforms where long and int are
32 bits, these overflow to negative values or wrap around, causing
heap buffer overflows, incorrect loop bounds, and undefined behavior.

dgif_lib.c:
- PixelCount: cast to (unsigned long) instead of signed (long) to
  prevent signed overflow UB on 32-bit platforms.
- LZW CodeSize: reject BitsPerPixel < 2 (CodeSize 0 and 1 create
  degenerate LZW states where ClearCode/EOFCode are unreachable).

egif_lib.c:
- PixelCount: same (unsigned long) fix as dgif_lib.c.
- EGifSpew: cast j to (size_t) in pointer arithmetic j * SavedWidth
  to prevent int*int overflow in raster offset computation.

gifalloc.c:
- GifApplyTranslation: use size_t for RasterSize and loop counter
  instead of int to prevent Height*Width overflow.
- GifMakeSavedImage: cast Height to (size_t) in reallocarray nmemb
  and memcpy size to prevent int*int overflow before reaching
  reallocarray's internal overflow check.

quantize.c:
- Compute PixelCount as (size_t)Width * Height with division-based
  overflow guard.  Use size_t loop counter for pixel iteration.
- Replace signed (long) cast of pixel count with (unsigned long).

All 51 regression tests pass.  Compiles with zero warnings under
-Wall -O2.

+++ egif_lib.c
@@ -448,7 +448,8 @@ int EGifPutImageDesc(GifFileType *GifFile, const int Left, const int Top,
 
 	/* Mark this file as has screen descriptor: */
 	Private->FileState |= FILE_STATE_IMAGE;
-	Private->PixelCount = (long)Width * (long)Height;
+	Private->PixelCount =
+	    (unsigned long)Width * (unsigned long)Height;
 
 	/* Reset compress algorithm parameters. */
 	(void)EGifSetupCompress(GifFile);
@@ -1154,7 +1155,7 @@ int EGifSpew(GifFileType *GifFileOut, int *ErrorCode) {
 				     j += InterlacedJumps[k]) {
 					if (EGifPutLine(
 					        GifFileOut,
-					        sp->RasterBits + j * SavedWidth,
+					        sp->RasterBits + (size_t)j * SavedWidth,
 					        SavedWidth) == GIF_ERROR) {
 						status = GIF_ERROR;
 						err = GifFileOut->Error;
@@ -1165,7 +1166,7 @@ int EGifSpew(GifFileType *GifFileOut, int *ErrorCode) {
 		} else {
 			for (j = 0; j < SavedHeight; j++) {
 				if (EGifPutLine(GifFileOut,
-				                sp->RasterBits + j * SavedWidth,
+				                sp->RasterBits + (size_t)j * SavedWidth,
 				                SavedWidth) == GIF_ERROR) {
 					status = GIF_ERROR;
 					err = GifFileOut->Error;
